requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pdfplumber==0.11.4
PyMuPDF==1.24.13
pdfminer.six==20231228
pytesseract==0.3.13
Pillow==11.0.0
langchain==0.3.7
langchain-community==0.3.5
sentence-transformers==3.3.1
chromadb==0.5.20
httpx==0.27.2
python-dotenv==1.0.1
